move browse.go
